Fix WebVTT Parser to disallow negative percentages https://bugs.webkit.org/show_bug.cgi?id=263640
Reviewed by Eric Carlson.

This patch addresses an issue in the WebVTT parser where it incorrectly
allowed negative percentage values for region and viewport anchors.
According to the WebVTT specification, percentage values must be in the
range of 0 to 100, and negative percentages are not valid. This patch
ensures that the parser correctly identifies negative percentages
as invalid values.

* LayoutTests/imported/w3c/web-platform-tests/webvtt/parsing/file-parsing/tests/regions-regionanchor-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/webvtt/parsing/file-parsing/tests/regions-viewportanchor-expected.txt:
* Source/WebCore/html/track/WebVTTParser.cpp:
(WebCore::WebVTTParser::parseFloatPercentageValue):

Canonical link: https://commits.webkit.org/269882@main