[WebGPU] Using the stencil reference value as the stencil clear value fails when the reference value is greater than 255 https://bugs.webkit.org/show_bug.cgi?id=263812 <radar://117610399>
Reviewed by Tadeu Zagallo.

The stencil value is masked to the max value in the stencil buffer.

All of our stencil formats are 8 bits and that is not likely to change
as stencil formats have not changed in more than ten years, so mask with 255.

* LayoutTests/http/tests/webgpu/webgpu/api/operation/render_pass/clear_value-expected.txt:
* Source/WebGPU/WebGPU/RenderPassEncoder.mm:
(WebGPU::RenderPassEncoder::setStencilReference):

Canonical link: https://commits.webkit.org/269957@main