RenderVideo::videoBox crashes when intrinsic size is zero https://bugs.webkit.org/show_bug.cgi?id=263990 <radar://116303559>
Reviewed by Alan Baradlay.

LayoutSize::fitToAspectRatio(aspectRatio, ) assumes that aspectRatio is
non-empty as it divides by aspectRatio.height() and aspectRatio.width().

When either are zero, this would result in a floating point exception due to
division by zero.

It's not clear we should add this check to fitToAspectRatio() and based on where
fitToAspectRatio is called, it seems more appropriate to check before the call site.

* Source/WebCore/rendering/RenderVideo.cpp:
(WebCore::RenderVideo::videoBox const):
Ensure that intrinsicSize is not empty.

Canonical link: https://commits.webkit.org/270032@main