Web Inspector: Font Panel: CSS font property values marked !important don't get overridden when using the interactive editing controls. https://bugs.webkit.org/show_bug.cgi?id=258975
Review by NOBODY (OOPS!).

Fix the font-weight slider as the !important declaration causes the slider to to not slide and be overridden.

Address the case of having multiple of the same property of multiple font-weights and a !important declaration in the inline style.

FontStyles.js, writeFontVariation(): check if the target property exists, check if the existing target property contains !important,
check if the existing property is font-weight, update the value as !important if not already,
create the font-weight property if it doesn't exist, and set the CSSProperty raw value without marking it as !important.

Add FixMe for font-weight being renamed to font-variation-settings under style attributes.
Remove FixMe in CSSProperty.js, as there are intentional calls to the CSSProperty.value getter in use.

* Source/WebInspectorUI/UserInterface/Models/CSSProperty.js:
* Source/WebInspectorUI/UserInterface/Models/FontStyles.js:
(WI.FontStyles.prototype.writeFontVariation):