Remove Quirk shouldDisableContentChangeObserver() for YouTube.com https://bugs.webkit.org/show_bug.cgi?id=264234 rdar://117975999
Reviewed by Tim Horton.

This removes the Quirk which had been set for YouTube in the context
of the iPad Pro, where people could not click on thumbnails on
search and results page.
https://bugs.webkit.org/show_bug.cgi?id=249740
it was then later extended to cover the full domain of YouTube.
https://bugs.webkit.org/show_bug.cgi?id=263789
YouTube team fixed the issue on their side. It removes the need for
this specific Quirk.

We probably want to take another look at
Quirks::shouldDisableContentChangeObserverTouchEventAdjustment()
and if it is still needed, but to minimize the effect of this fix,
it will be handled in another PR if necessary.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldDisableContentChangeObserver const): Deleted.
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/ios/ContentChangeObserver.cpp:
(WebCore::ContentChangeObserver::isContentChangeObserverEnabled):
* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::handleSyntheticClick):
(WebKit::WebPage::completeSyntheticClick):

Canonical link: https://commits.webkit.org/270256@main